+++ packages/save-and-share-bar/src/save-and-share-bar.js
@@ -10,6 +10,7 @@ import UserState from "@times-components/user-state";
 import { SectionContext } from "@times-components/context";
 import { SaveStar } from "@times-components/ts-components";
 
+import { checkForSymphonyExperiment } from "@times-components/utils";
 import getTokenisedArticleUrlApi from "./get-tokenised-article-url-api";
 import withTrackEvents from "./tracking/with-track-events";
 import SharingApiUrls from "./constants";
@@ -18,6 +19,8 @@ import styles from "./styles";
 import {
   SaveAndShareBarContainer,
   ShareButtonContainer,
+  ShareButtonHighlightContainer,
+  ShareButtonHighlight,
   OutlineButton,
   Popover,
   PopoverHeader,
@@ -48,6 +51,7 @@ function SaveAndShareBar(props) {
   const [popoverOpen, setPopoverOpen] = React.useState(false);
   const [windowHeight, setWindowHeight] = React.useState(null);
   const [windowWidth, setWindowWidth] = React.useState(null);
+  const [shareButtonClicked, setShareButtonClicked] = React.useState(false);
 
   const barRef = React.useRef();
   const shareBtnRef = React.useRef();
@@ -105,6 +109,13 @@ function SaveAndShareBar(props) {
     setPopoverOpen(prev => !prev);
   };
 
+  const handleClick = e => {
+    e.preventDefault();
+    setShareButtonClicked(true);
+    togglePopover();
+    window.localStorage.setItem("hasShareButtonHighlightBeenDismissed", true);
+  };
+
   function copyToClipboard(e) {
     const { onCopyLink } = props;
     e.preventDefault();
@@ -113,6 +124,15 @@ function SaveAndShareBar(props) {
     onCopyLink();
   }
 
+  const isSymphonyExperiment = checkForSymphonyExperiment();
+  const hasShareButtonHighlightBeenDismissed = window.localStorage.getItem(
+    "hasShareButtonHighlightBeenDismissed"
+  );
+  const showShareButtonHighlight =
+    isSymphonyExperiment &&
+    !shareButtonClicked &&
+    !hasShareButtonHighlightBeenDismissed;
+
   return (
     <SaveAndShareBarContainer data-testid="save-and-share-bar" ref={barRef}>
       {sharingEnabled && (
@@ -120,11 +140,16 @@ function SaveAndShareBar(props) {
           <OutlineButton
             ref={shareBtnRef}
             isPopoverOpen={popoverOpen}
-            onClick={togglePopover}
+            onClick={handleClick}
           >
             <ShareIcon height={14} width={14} />
             Share
           </OutlineButton>
+          {showShareButtonHighlight && (
+            <ShareButtonHighlightContainer data-testId="share-button-highlight">
+              <ShareButtonHighlight />
+            </ShareButtonHighlightContainer>
+          )}
           <Popover
             ref={popoverRef}
             position={getPosition()}