+++ packages/ts-components/src/components/social-embed/__tests__/SocialVendor.test.tsx
@@ -23,6 +23,12 @@ jest.mock('../components/TiktokComponent', () => ({
   ))
 }));
 
+jest.mock('../components/InstagramComponent', () => ({
+  Instagram: jest.fn(() => (
+    <div data-testid="instagram-component">Instagram Component</div>
+  ))
+}));
+
 describe('Vendor Component', () => {
   const renderVendor = (vendorName: VendorName, url: string) =>
     render(<Vendor vendorName={vendorName} url={url} />);
@@ -35,6 +41,11 @@ describe('Vendor Component', () => {
   it('renders the Youtube component when vendorName is "youtube"', () => {
     renderVendor('youtube' as VendorName, 'https://youtube.com/some/video');
     expect(screen.getByTestId('youtube-component')).toBeInTheDocument();
+  });
+
+  it('renders the Instagram component when vendorName is "instagram"', () => {
+    renderVendor('instagram' as VendorName, 'https://instagram.com/some/video');
+    expect(screen.getByTestId('instagram-component')).toBeInTheDocument();
   });
 
   it('renders the TikTok component when vendorName is "tiktok"', () => {