+++ packages/utils/src/index.js
@@ -23,6 +23,12 @@ export { default as TcFlatList } from "./tc-flat-list";
 export { default as TcText } from "./tc-text";
 export { default as TcTextLink } from "./tc-text-link";
 export { default as TcScrollView } from "./tc-scroll-view";
+export { debounce } from "./debounce";
 export {
   default as CanShowPuzzleSidebar
 } from "./article-puzzle-sidebar-whitelist";
+export { default as getCookieValue } from "./get-cookie-value";
+export { default as getBase64CookieValue } from "./get-base64-cookie-value";
+export {
+  default as checkForSymphonyExperiment
+} from "./check-for-symphony-experiment";