+++ packages/link/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.17.35](https://github.com/newsuk/times-components/compare/@times-components/link@3.17.34...@times-components/link@3.17.35) (2025-03-12)
+
+**Note:** Version bump only for package @times-components/link
+
+
+
+
+
 ## [3.17.34](https://github.com/newsuk/times-components/compare/@times-components/link@3.17.33...@times-components/link@3.17.34) (2024-07-02)
 
 **Note:** Version bump only for package @times-components/link