+++ packages/utils/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [6.27.0](https://github.com/newsuk/times-components/compare/@times-components/utils@6.26.0...@times-components/utils@6.27.0) (2025-03-12)
+
+
+### Features
+
+* **TMG-3003:** Add blue dot to share button ([#4079](https://github.com/newsuk/times-components/issues/4079)) ([d0e76b3](https://github.com/newsuk/times-components/commit/d0e76b38891228624bf0415828b1179c4c2c596e))
+
+
+
+
+
 # [6.26.0](https://github.com/newsuk/times-components/compare/@times-components/utils@6.25.1...@times-components/utils@6.26.0) (2024-07-02)